normalizing the memory extraction interfaces
